simplify sidechain obj creation
